server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
google-generativeai==0.8.3
chromadb==0.4.18
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.6
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.112.0
aiofiles==23.2.1
aiohttp==3.9.1
pytz==2025.2